Move files to a more its respective folders
